pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-webtester"
version = "0.1.0"
description = "AI-powered web application testing framework with browser automation and real-time monitoring"
authors = [{name = "AI WebTester", email = "contact@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12,<0.15",
    "playwright>=1.46,<1.49",
    "fastapi>=0.111,<0.116",
    "uvicorn[standard]>=0.27,<0.32",
    "websockets>=12,<15",
    "pyyaml>=6,<7",
    "jinja2>=3.1,<3.2",
    "beautifulsoup4>=4.12,<4.13",
    "aiofiles>=23,<24",
    "openai>=1.66,<2",
    "pydantic>=2.5,<3",
    "python-multipart>=0.0.6,<1",
    "backoff>=2.2,<3",
    "faker>=37.6.0",
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.8",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0,<8",
    "pytest-asyncio>=0.21.0,<1",
    "black>=23.0.0,<25",
    "ruff>=0.1,<1",
    "mypy>=1.5,<2",
    "pre-commit>=3.0,<4"
]

[project.scripts]
ai-webtester = "cli.main:app"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.hatch.build.targets.wheel]
packages = ["orchestrator", "cli", "browser", "evidence", "mock_app", "utils", "providers", "data_gen"]

[tool.hatch.build.targets.wheel.force-include]
"configs" = "configs"
"mock_app/templates" = "mock_app/templates"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
